RBAC Implemented

+++ app.py
@@ -25,7 +25,7 @@ def signup(data: UserCreate, db: Session = Depends(get_db)):
     if db.query(User).filter(User.username == data.username).first():
         raise HTTPException(status_code=400, detail="Username already exists")
     
-    user = User(username=data.username, password=hash_password(data.password))
+    user = User(username=data.username, role="student", password=hash_password(data.password))
     db.add(user)
     db.commit()
     db.refresh(user)
@@ -38,13 +38,16 @@ def login(data: UserLogin, db: Session = Depends(get_db)):
     if not user or not verify_password(data.password, user.password):
         raise HTTPException(status_code=401, detail="Invalid credentials")
 
-    token = create_access_token({"sub": user.username})
+    token = create_access_token({"sub": user.username, "role": user.role})
 
     return {"access_token": token}
 
 @app.post("/students", status_code=201, response_model=StudentResponse)
-def create_student(data: StudentCreate, db: Session = Depends(get_db)):
+def create_student( data: StudentCreate, user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
     try:
+        required_fields = ["admin", "teacher"]
+        if user["role"] not in required_fields:
+            raise HTTPException(status_code=403, detail="You do not have permission to perform this action!")
         student = Student(**data.model_dump())
         return service.add_student(db, student)
 
@@ -52,15 +55,15 @@ def create_student(data: StudentCreate, db: Session = Depends(get_db)):
         raise HTTPException(status_code=400, detail=str(e))
     
 @app.get("/students", response_model=list[StudentResponse])
-def get_students(user: str = Depends(get_current_user),db: Session = Depends(get_db), query: StudentQuery = Depends()):
+def get_students(user: dict = Depends(get_current_user),db: Session = Depends(get_db), query: StudentQuery = Depends()):
     return service.get_all_students(db, query)
 
 @app.get("/students/count")
-def get_student_count(db: Session = Depends(get_db)):
+def get_student_count(user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
     return {"count": service.count_students(db)}
 
 @app.get("/students/{student_id}", response_model=StudentResponse)
-def get_student(student_id: str, db: Session = Depends(get_db)):
+def get_student(student_id: str, user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
     try:
         student = service.find_student(db, student_id)
         if not student:
@@ -71,8 +74,11 @@ def get_student(student_id: str, db: Session = Depends(get_db)):
         raise HTTPException(status_code=404, detail=str(e))
 
 @app.delete("/students/{student_id}")
-def delete_student(student_id: str, db: Session = Depends(get_db)):
+def delete_student(student_id: str, user: dict = Depends(get_current_user),  db: Session = Depends(get_db)):
     try:
+        required_role = "admin"
+        if user["role"] != required_role:
+            raise HTTPException(status_code=403, detail="You do not have permission to perform this action!")
         service.delete_student(db, student_id)
         return {"message": "Deleted student successfully!"}
     
@@ -80,8 +86,11 @@ def delete_student(student_id: str, db: Session = Depends(get_db)):
         raise HTTPException(status_code=404, detail=str(e))
 
 @app.put("/students/{student_id}", status_code=200, response_model=StudentResponse)
-def update_student(student_id: str, data: StudentUpdate, db: Session = Depends(get_db)):
+def update_student(student_id: str, data: StudentUpdate, user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
     try:
+        required_role = ["admin", "teacher"]
+        if user["role"] not in required_role:
+            raise HTTPException(status_code=403, detail="You do not have permission to perform this action!")
         return service.update_student(
             db,
             student_id, 
@@ -93,8 +102,11 @@ def update_student(student_id: str, data: StudentUpdate, db: Session = Depends(g
         raise HTTPException(status_code=404, detail=str(e))
     
 @app.post("/courses", response_model=CourseResponse)
-def create_course(data: CourseCreate, db: Session = Depends(get_db)):
+def create_course(data: CourseCreate, user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
     try:
+        required_role = ["admin", "teacher"]
+        if user["role"] not in required_role:
+            raise HTTPException(status_code=403, detail="You do not have permission to perform this action!")
         course = Course(**data.model_dump())
         return CourseService().add_course(db, course)
